+++ CHANGES.md
@@ -6,6 +6,7 @@
 
 - [#054] Fixed version displayed as `2.0.0.0` instead of `2.0.0` when config file is missing
 - [#053] Added support for Symfony's UX-Icons: https://ux.symfony.com/icons
+- [#051] Added support for `config/packages/disco_toolbar.yaml` config location
 
 ## 2.0.0 (2026-01-30)